+++ docs/src/docs/asciidoc/core-concepts.adoc
@@ -479,8 +479,8 @@ if (event instanceof TickEvent) {
 
 TamboUI tries to enforce a clear, consistent exception hierarchy for framework errors.
 
-* All framework exceptions extend `TamboUIException` (`dev.tamboui.errors.TamboUIException`). This is a `RuntimeException`.
-* Terminal I/O errors are wrapped in `RuntimeIOException` (`dev.tamboui.errors.RuntimeIOException`), to avoid having to catch `IOException` which often are non-recoverable.
+* All framework exceptions extend `TamboUIException` (`dev.tamboui.error.TamboUIException`). This is a `RuntimeException`.
+* Terminal I/O errors are wrapped in `RuntimeIOException` (`dev.tamboui.error.RuntimeIOException`), to avoid having to catch `IOException` which often are non-recoverable.
 * Backend (non-I/O) errors use `BackendException` (`dev.tamboui.terminal.BackendException`).
 * TUI framework errors use `TuiException` (`dev.tamboui.tui.TuiException`).
 * For invalid parameters or state, Java standard exceptions such as `IllegalArgumentException` or `IllegalStateException` are used.
