+++ tamboui-core/src/main/java/dev/tamboui/error/TerminalIOException.java
@@ -2,7 +2,11 @@
  * Copyright TamboUI Contributors
  * SPDX-License-Identifier: MIT
  */
-package dev.tamboui.errors;
+package dev.tamboui.error;
+
+import java.io.IOException;
+
+import dev.tamboui.error.TamboUIException;
 
 /**
  * Exception thrown when a terminal I/O operation fails.
@@ -11,29 +15,26 @@ package dev.tamboui.errors;
  * such as reading from or writing to the terminal, setting terminal attributes,
  * or querying terminal state.
  *
- * NOTE: This is similar to java.io.UncheckedExcpetion, but does not share name to avoid
- * ambiguity
- *
  * @see TamboUIException
  */
-public class RuntimeIOException extends TamboUIException {
+public class TerminalIOException extends TamboUIException {
 
     /**
-     * Creates a new runtime/unchecked IO exception with the given message.
+     * Creates a new terminal I/O exception with the given message.
      *
      * @param message the error message
      */
-    public RuntimeIOException(String message) {
+    public TerminalIOException(String message) {
         super(message);
     }
 
     /**
-     * Creates a new runtime/unchecked I/O exception with the given message and cause.
+     * Creates a new terminal I/O exception with the given message and cause.
      *
      * @param message the error message
      * @param cause   the underlying IOException
      */
-    public RuntimeIOException(String message, Throwable cause) {
+    public TerminalIOException(String message, Throwable cause) {
         super(message, cause);
     }
