+++ tamboui-core/src/main/java/dev/tamboui/internal/record/AnsiTerminalCapture.java
@@ -12,7 +12,7 @@ import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.List;
 
-import dev.tamboui.errors.RuntimeIOException;
+import dev.tamboui.error.RuntimeIOException;
 
 /**
  * Captures raw bytes written to System.out for recording inline demos.
