+++ tamboui-core/src/main/java/dev/tamboui/terminal/BackendFactory.java
@@ -4,7 +4,7 @@
  */
 package dev.tamboui.terminal;
 
-import dev.tamboui.errors.TamboUIException;
+import dev.tamboui.error.TamboUIException;
 import dev.tamboui.internal.record.RecordingBackend;
 import dev.tamboui.internal.record.RecordingConfig;
 import dev.tamboui.util.SafeServiceLoader;
