+++ tamboui-core/src/main/java11/module-info.java
@@ -19,7 +19,7 @@ module dev.tamboui.core {
     exports dev.tamboui.layout.grid;
     exports dev.tamboui.layout.stack;
     exports dev.tamboui.widget;
-    exports dev.tamboui.errors;
+    exports dev.tamboui.error;
 
     uses dev.tamboui.terminal.BackendProvider;
 }
