+++ tamboui-panama-backend/src/main/java/dev/tamboui/backend/panama/unix/UnixTerminal.java
@@ -6,7 +6,7 @@ package dev.tamboui.backend.panama.unix;
 
 import dev.tamboui.terminal.BackendException;
 import dev.tamboui.backend.panama.PlatformTerminal;
-import dev.tamboui.errors.RuntimeIOException;
+import dev.tamboui.error.RuntimeIOException;
 import dev.tamboui.layout.Size;
 
 import java.io.IOException;
