+++ tamboui-toolkit/src/main/java/dev/tamboui/toolkit/element/StyledElement.java
@@ -13,8 +13,8 @@ import dev.tamboui.toolkit.event.EventResult;
 import dev.tamboui.toolkit.event.KeyEventHandler;
 import dev.tamboui.toolkit.event.MouseEventHandler;
 import dev.tamboui.toolkit.id.IdGenerator;
-import dev.tamboui.tui.TuiException;
 import dev.tamboui.tui.bindings.ActionHandler;
+import dev.tamboui.tui.error.TuiException;
 import dev.tamboui.layout.Constraint;
 import dev.tamboui.layout.Rect;
 import dev.tamboui.style.Color;
