+++ tamboui-toolkit/src/main/java/dev/tamboui/toolkit/elements/ListElement.java
@@ -14,8 +14,8 @@ import dev.tamboui.layout.Rect;
 import dev.tamboui.style.Color;
 import dev.tamboui.style.Style;
 import dev.tamboui.terminal.Frame;
-import dev.tamboui.tui.TuiException;
 import dev.tamboui.tui.bindings.Actions;
+import dev.tamboui.tui.error.TuiException;
 import dev.tamboui.tui.event.KeyEvent;
 import dev.tamboui.tui.event.MouseEvent;
 import dev.tamboui.tui.event.MouseEventKind;
