--- tamboui-tui/src/main/java/dev/tamboui/tui/TuiException.java
@@ -1,17 +1,0 @@
-package dev.tamboui.tui;
-
-import dev.tamboui.errors.TamboUIException;
-
-/**
- * Exception for TUI framework related errors.
- */
-public class TuiException extends TamboUIException {
-
-    public TuiException(String message) {
-        super(message);
-    }
-
-    public TuiException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
