+++ tamboui-widgets/src/test/java/dev/tamboui/widgets/block/BlockTest.java
@@ -5,7 +5,7 @@
 package dev.tamboui.widgets.block;
 
 import dev.tamboui.buffer.Buffer;
-import dev.tamboui.errors.RuntimeIOException;
+import dev.tamboui.error.RuntimeIOException;
 import dev.tamboui.layout.Alignment;
 import dev.tamboui.layout.Padding;
 import dev.tamboui.layout.Rect;
